[Server] - Updated Files Path to watch easily
