+++ server/src/app/routes/graphql/mutation/index.ts
@@ -2,13 +2,10 @@ import { GraphQLNonNull, GraphQLObjectType } from 'graphql';
 import { GraphQLBoolean, GraphQLInt, GraphQLString } from 'graphql/type/scalars';
 import { BookType, DivisionType, UserType } from '../type/index';
 
-import { BookRepository } from '../../../repository/book/BookRepository';
-import { DivisionRepository } from '../../../repository/division/DivisionRepository';
-
-import { AuthorityMutation } from './authority';
-import { BookMutation } from './book';
-import { DivisionMutation } from './division';
-import { UserMutation } from './user';
+import { AuthorityMutation } from './AuthorityMutation';
+import { BookMutation } from './BookMutation';
+import { DivisionMutation } from './DivisionMutation';
+import { UserMutation } from './UserMutation';
 
 const mutation = new GraphQLObjectType({
   name: 'Mutation',
