+++ server/src/app/services/BookService.ts
