+++ server/src/config/develop.ts
@@ -4,5 +4,5 @@ export default {
 
   // Oauth2
   GITHUB_CLIENT_ID: '05096350eaddf80dbd34',
-  GITHUB_CLIENT_SECRET_ID: 'a2752bd252c43ef3d93f6ec4ac29533f8ff0a087',
+  GITHUB_CLIENT_SECRET_ID: '773371ce78f5f2578873f75f202759c0b28b85e0',
 };
